was refactored and modified code
